pyproject.toml
[tool.poetry]
name = "tubesage"
version = "0.1.0"
description = "LLM that understands a video transcript and answers questions about it"
authors = ["Gustavo Lima <correia.gustavol@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
youtube-transcript-api = "^0.6.2"
flask = "^3.0.3"
langchain-community = "^0.2.10"
langchain-text-splitters = "^0.2.2"
langchain-chroma = "^0.1.2"
langchain-ollama = "^0.1.1"
flask-restful = "^0.3.10"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
